PIA/requirements.txt
streamlit==1.25.0  
requests==2.31.0  
python-dotenv==1.0.0  
azure-cosmos==4.3.0